CLOUDP-349087: Block simultaneous TLS disabling and scaling for ReplicaSets (#549)
# Summary

We have code in the replica set controller dedicated to handling a
scenario where the user disable TLS, and ascale their cluster at the
same time. The test for this behaviour was broken. Because there were
two functions with the name `test_tls_is_disabled_and_scaled_up`, so
pytest would run only one of them.

When I fixed the test, it failed:
https://spruce.mongodb.com/task/mongodb_kubernetes_e2e_mdb_kind_ubi_cloudqa_e2e_disable_tls_scale_up_patch_dfb9424e9b34ddd048a725a9988114ca4032f9bf_68de79b2fccf070007a2c51e_25_10_02_13_10_12/tests?execution=4&sorts=STATUS%3AASC

Which means the code to handle it was incorrect. My opinion is that it
is better to block this change, rather than introducing complexity to
handle it. It is not a common use case.

A changelog entry was added. Question: should it be explicitly mentioned
in our public documentation ?

Old related PR (2021):
https://github.com/10gen/ops-manager-kubernetes/pull/1444
And ticket: https://jira.mongodb.org/browse/CLOUDP-80768

## Proof of Work

Unit and e2e test for blocking the operation should pass.
No regression due to the change in `updateOmDeploymentRs` should be
observed.

## Checklist

- [X] Have you linked a jira ticket and/or is the ticket in the title?
- [X] Have you checked whether your jira ticket required DOCSP changes?
- [X] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details