Remove `def _helm_chart_dir` that used undefined env var (#535)
# Summary

In our source file we had a function 

```
def _helm_chart_dir(default: Optional[str] = "helm_chart") -> str:
    return os.environ.get("HELM_CHART_DIR", default)
```
The env var `HELM_CHART_DIR` is not set in our project anywhere that
makes this function not needed. That's why this PR removes the function.

From the function def we can see that it would always return what was
passed to this function, that's why at the caller we are removing the
function call and just using the argument.

## Proof of Work

CI in the PR.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details