Use const for `helm_chart` local chart dir (#532)
# Summary

The local helm chart directory (`helm_chart`) that we use to install the
operator is being used at a lot of places and we were using the dir name
in line without using the defined constant. This would make it harder to
use any other dir as helm chart dir/URL, that we are going to do when we
start installing helm chart from OCI registry.
This PR changes that and tries to use the const everywhere we used the
local chart dir (`helm_chart`).

edit: To make this work I had to move all constants from
`tests.conftest` to `tests.constants` file to remove circular
dependency.

## Proof of Work

CI in the PR.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details

---------

Co-authored-by: Maciej Karaś <6159874+MaciejKaras@users.noreply.github.com>
Co-authored-by: Maciej Karaś <maciej.karas@mongodb.com>