CLOUDP-334392 - Add Delve sidecar for debugging the agent (#596)
# Summary

This PR removes any code that downloads go & delve at runtime from the
agent-launcher script. Instead of enabling debugging of the agent from
the script, this is now enabled via the operator. When setting
`MDB_AGENT_DEBUG` to true in the private context, the operator will add
a sidecar which runs delve by attaching to the agent's process.

Instructions for using can be found in our
[wiki](https://wiki.corp.mongodb.com/spaces/MMS/pages/427469143/Debugging+a+running+Automation+Agent+in+a+cluster).

## Proof of Work

This is a local development improvement. It can be checked by adding the
env variable to the context and seeing the delve sidecar running.
```
~ % kubectl get pods my-mc-replicaset-0-0 -o jsonpath='{.spec.initContainers[*].name}'
delve-sidecar mongodb-kubernetes-init-database istio-init
```

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details