CLOUDP-353899: Services generation (#619)
# Summary

Note: This is a branch I was working on before we paused the MDBMC
project. I'd like to merge the work in the main feature branch since I
was done with it (tests were green). It will make it much easier to
resume work on the main feature branch in the future if we decide to
continue it.

Generates the necessary services for multi cluster, like in mongodb
multi replica set controller

## Proof of Work

Passing unit test + basic `e2e_multi_cluster_new_replica_set_scale_up`
e2e test
Full E2E tests suite will be added later in the feature branch

## Checklist

- [X] Have you linked a jira ticket and/or is the ticket in the title?
- [X] Have you checked whether your jira ticket required DOCSP changes?