CLOUDP-328217: Automation agent password secret (#566)
# Summary

If a deployment is moved to a different project, the automation agent
password will be re-generated, triggering a password change in the
automation plan.

This will cause a deadlock in a sharded cluster due to the multiple
components requiring automation. However, it will not cause issues in
replicasets.

This is a blocker for migrating projects in sharded deployments.

## Proof of Work

For SCRAM (the only auth mechanism who re-generates a pwd), we now save
the automation agent's password in a secret. During migration, the
stored secret is utilized to preserve the password ,ensuring project
migration possible.

## Observed problems
For LDAP (Sharded + Replica), the following tests are failing, even
though the only modification made is updating the MongoDB resource's
project reference, with no other changes applied. To help further
investigation, I have commented out certain code in the tests (which can
make them fail) so the issue can be consistently reproduced. While the
deployment returns to the "running" state, the users are missing from
the automation configuration.

Additionally, certain flaky parts in the SHA1 and SHA256 tests, as well
as the LDAP and X509 project switch tests, have been commented
out/disabled due to the lack of support for project migrations. These
tests should be activated once project migration functionality has been
implemented.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details