[CLOUDP-361632] Add information about latest OM and agent in release.json (#622)
# Summary

We have been getting some request from the customers where they wanted
to know the combination of the container images that they require to run
a specific version of operator and Mongo DB instance. As part of that
work we are going to add a new file (release_info.json) as the assets in
the GitHub release.
That file is going to have details about the latest released images.
These images are going to be the images that MCK publishes as well as
the latest OM and Agent version mappings. To enable that, as part of
this PR we are adding a new field in the `release.json` that has the
mapping between OM version and agent version of latest OM releases. This
new field in the `release.json` is going to have the opsManaget to agent
image version mapping for every major opsmanager version.
We are adding this information in the `release.json` so that it can
easily be used later (in next PR), to create the release_info.json file.
This PR also updated the precommit hook so that when new version of OM
is released, the newly added field of release.json would properly be
updated with newly release OM/agent.

## Proof of Work

Manual: 
- Mimic new release of OM by adding a new version (8.0.17) to
`supportedImages.ops-manager.versions` and respective entry in
`mongodb-agent.opsManagerMapping`. Run `make precommit` and verify that
the `supportedImages.latestOpsManagerAgentMapping` has been updated for
major version 8 with updated versions.
- Mimic the new major version release of OM by adding new version
(9.0.1) to `supportedImages.ops-manager.versions` and respective entry
in `mongodb-agent.opsManagerMapping`. Run `make precommit` and verify
that a new entry for major version 9 has been added to
`supportedImages.latestOpsManagerAgentMapping`.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you added changelog file? (will show up later)
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details