[CLOUDP-363995] Rename `repositories` to `repository` and add a secondary field in `build_info.json` (#633)
# Summary

`build_info.json` has a field called `repositories` for images. If more
than one repos (`agent` is an example) are mentioned for an image it
becomes confusing as to which one should be used. Below is an example

```
"repositories": ["quay.io/mongodb/mongodb-agent-ubi", "quay.io/mongodb/mongodb-agent"],
```

For example if we have to add this information in
`release_info_<mck-version>.json` which repo should be added for agent
image.
That's why this PR changes the format slightly and renames
`repositories` to `repository` and adds a new field
`secondary_repositories`.
`repository` would be the main repo and other repos would be in the new
field secondary_repositories.

## Proof of Work

Unit test passes and successful CI.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details