CLOUDP-355710 Fix deletion of external roles (#625)
# Summary

This PR fixes an issue where roles added to an OM project outside the
operator (through UI or the API) are overwritten by the operator every
reconciliation. This is now consistent with the user behaviour, where
users can be defined in the UI and will not be removed.

To accomplish this, we are now tracking the roles we configure with the
operator in an annotation, and state configmap (where applicable). This
was necessary due to possible usage of the `ClusterMongoDBRole`
resource. If the reference to a role resource is removed at the same
time as removing the resource itself, then it is not possible to
determine which role needs to be removed from the automation config
without knowing the previous set of roles.

## Proof of Work

Unit tests have been added to verify that the `ensureRoles` method which
is reused across all controllers behaves correctly. Additionally, the
unit tests verify that each controller keeps track of the previously
configured roles in an annotation.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details