[CLOUDP-349630] Don't generate scramshacreds if `MongoDBUser`'s password is not changed (#628)
# Summary

When the reconciliation happens for the `MongoDBUser` resource (because
of controller restart or because of any other reason), for `SCRAM`
authentication mechanism, the scram sha creds are generated and then
they are set in the automation config which results into another version
getting created for Automation Config.
This workflow can be improved to generated the scram sha creds only when
the password of the user is changed. If we do that we will not be
changing the user's scram creds and the automation config will not be
updated.
To achieve this we are checking if the scram sha creds that are in
automation config for a user, match with the scram sha creds generated
with the stored salt and new password. If the scram sha creds match it
means that the password for the user is not changed and we don't have to
generate the new scram sha creds, otherwise new scram sha creds will be
generated.

## Proof of Work

Unit test:

```
~/work/opensource/mongodb-kubernetes/controllers/operator/authentication (dont-rotate-scram-creds-mdbu) » go test -timeout 30s -run ^Test_isPasswordChanged$ -v
=== RUN   Test_isPasswordChanged
--- PASS: Test_isPasswordChanged (0.01s)
PASS
ok      github.com/mongodb/mongodb-kubernetes/controllers/operator/authentication       0.472s
```

I am yet to do some more tests to actually be confident that the
automation config is not changed when reconciliation happens.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details