Fix the step publishing the draft release (#650)
We currently set `prerelease` on the draft release in the
`crate_draft_release_notes` job.

Later in the `publish_release` job we try to publish it, but it fails
with the error: "Latest release cannot be draft or prerelease."

With this change we will not be setting `prerelease`
when creating a draft.